remove currentTenantSlice
