+++ src/modules/share/states.ts
@@ -1,0 +1,41 @@
+import { atom, selector } from 'recoil';
+
+export type Tenant = {
+  id: string;
+  name: string;
+  logo?: string;
+  description?: string;
+  subdomain: string;
+  tenantId: string;
+  theme: string;
+  background1?: string;
+  background2?: string;
+  background3?: string;
+};
+
+export const currentTenantDataAtom = atom<Tenant>({
+  key: 'currentTenant',
+  default: {
+    id: '',
+    name: '',
+    subdomain: '',
+    tenantId: '',
+    theme: 'green',
+  },
+});
+
+export const currentTenantSubdomainSelector = selector({
+  key: 'currentTenantSubdomain',
+  get: ({ get }) => {
+    const tenant = get(currentTenantDataAtom);
+    return tenant?.subdomain;
+  },
+});
+
+export const currentTenantThemeSelector = selector({
+  key: 'currentTenantTheme',
+  get: ({ get }) => {
+    const tenant = get(currentTenantDataAtom);
+    return tenant?.theme;
+  },
+});
